Add an cast to avoid hiding type conversion
The operands are converted from integer to double and
the cast makes this explicit.

+++ cameraview/src/main/java/de/markusfisch/android/cameraview/widget/CameraView.java
@@ -274,7 +274,7 @@ public class CameraView extends FrameLayout {
 		Camera.Size bestSizeAspect = null;
 
 		for (Camera.Size size : sizes) {
-			double diff = Math.abs(size.height - height) +
+			double diff = (double) Math.abs(size.height - height) +
 					Math.abs(size.width - width);
 
 			if (diff < minDiff) {
